+++ packages/instabug_flutter/android/src/test/java/com/instabug/flutter/InstabugApiTest.java
@@ -779,4 +779,15 @@ public class InstabugApiTest {
         mInstabug.verify(() -> Instabug.setAutoMaskScreenshotsTypes(MaskingType.LABELS,MaskingType.MEDIA,MaskingType.TEXT_INPUTS,MaskingType.MASK_NOTHING));
     }
 
+
+    @Test
+    public void testGetNetworkBodyMaxSize() {
+        double expected = 10240;
+        InstabugPigeon.Result<Double> result = makeResult((actual) -> assertEquals((Double) expected, actual));
+
+        mockkObject(new InternalCore[]{InternalCore.INSTANCE}, false);
+        every(mockKMatcherScope -> InternalCore.INSTANCE.get_networkLogCharLimit()).returns((int) expected);
+
+        api.getNetworkBodyMaxSize(result);
+    }
 }