+++ packages/instabug_flutter/lib/src/utils/feature_flags_manager.dart
@@ -9,6 +9,8 @@ typedef OnW3CFeatureFlagChange = void Function(
   bool isW3cCaughtHeaderEnabled,
 );
 
+typedef OnNetworkBodyMaxSizeChangeCallback = void Function();
+
 class FeatureFlagsManager implements FeatureFlagsFlutterApi {
   // Access the singleton instance
   factory FeatureFlagsManager() {
@@ -22,6 +24,10 @@ class FeatureFlagsManager implements FeatureFlagsFlutterApi {
 
   // Host API instance
   static InstabugHostApi _host = InstabugHostApi();
+
+  // Callback for network body max size changes
+  static OnNetworkBodyMaxSizeChangeCallback?
+      _onNetworkBodyMaxSizeChangeCallback;
 
   /// @nodoc
   @visibleForTesting
@@ -38,10 +44,21 @@ class FeatureFlagsManager implements FeatureFlagsFlutterApi {
     // since it breaks the singleton pattern
   }
 
+  /// Sets the callback for network body max size changes
+  // ignore: avoid_setters_without_getters
+  set onNetworkBodyMaxSizeChangeCallback(
+    OnNetworkBodyMaxSizeChangeCallback callback,
+  ) {
+    _onNetworkBodyMaxSizeChangeCallback = callback;
+  }
+
   // Internal state flags
   bool _isAndroidW3CExternalTraceID = false;
   bool _isAndroidW3CExternalGeneratedHeader = false;
   bool _isAndroidW3CCaughtHeader = false;
+  int _networkBodyMaxSize = 0;
+
+  int get networkBodyMaxSize => _networkBodyMaxSize;
 
   Future<W3cFeatureFlags> getW3CFeatureFlagsHeader() async {
     if (IBGBuildInfo.instance.isAndroid) {
@@ -64,9 +81,10 @@ class FeatureFlagsManager implements FeatureFlagsFlutterApi {
     );
   }
 
-  Future<void> registerW3CFlagsListener() async {
+  Future<void> registerFeatureFlagsListener() async {
     FeatureFlagsFlutterApi.setup(this); // Use 'this' instead of _instance
 
+    // W3C Feature Flags
     final featureFlags = await _host.isW3CFeatureFlagsEnabled();
     _isAndroidW3CCaughtHeader =
         featureFlags['isW3cCaughtHeaderEnabled'] ?? false;
@@ -74,6 +92,10 @@ class FeatureFlagsManager implements FeatureFlagsFlutterApi {
         featureFlags['isW3cExternalTraceIDEnabled'] ?? false;
     _isAndroidW3CExternalGeneratedHeader =
         featureFlags['isW3cExternalGeneratedHeaderEnabled'] ?? false;
+
+    // Network Body Max Size
+    final networkBodyMaxSize = await _host.getNetworkBodyMaxSize();
+    _networkBodyMaxSize = networkBodyMaxSize?.toInt() ?? 0;
 
     return _host.registerFeatureFlagChangeListener();
   }
@@ -88,5 +110,11 @@ class FeatureFlagsManager implements FeatureFlagsFlutterApi {
     _isAndroidW3CCaughtHeader = isW3cCaughtHeaderEnabled;
     _isAndroidW3CExternalTraceID = isW3cExternalTraceIDEnabled;
     _isAndroidW3CExternalGeneratedHeader = isW3cExternalGeneratedHeaderEnabled;
+  }
+
+  @override
+  void onNetworkLogBodyMaxSizeChange(int networkBodyMaxSize) {
+    _networkBodyMaxSize = networkBodyMaxSize;
+    _onNetworkBodyMaxSizeChangeCallback?.call();
   }
 }