+++ packages/instabug_flutter/pigeons/instabug.api.dart
@@ -7,6 +7,8 @@ abstract class FeatureFlagsFlutterApi {
     bool isW3cExternalGeneratedHeaderEnabled,
     bool isW3cCaughtHeaderEnabled,
   );
+
+  void onNetworkLogBodyMaxSizeChange(int networkBodyMaxSize);
 }
 
 @HostApi()
@@ -108,6 +110,9 @@ abstract class InstabugHostApi {
   void willRedirectToStore();
 
   void setNetworkLogBodyEnabled(bool isEnabled);
+
+  @async
+  double? getNetworkBodyMaxSize();
 
   void setTheme(Map<String, Object> themeConfig);
   void setFullscreen(bool isEnabled);